fix: Add GET implementation of OCSP Server
